Added search functionality
A new route (/items/search/{search_term}) has been added to the API to allow for users on the site to easy look through the title of items.

For futureproofing, I decided to take the longer route and implement this on the back-end instead of filtering it easily on the client side, which isn't likely to scale well as the database grows in size.

I aim to further this functionality by allowing users to search by date and/or content.