New endpoint for determing whether a chamber is in session
New endpoint, "/info/session/{chamber}" will be used on the front-end to display whether the House/Senate is in session. On the backend, the program will automatically send recent releases from the House/Senate press galleries, and feed it into the Mixtral 8x7b large language model. It will then return when the next time each chamber is meeting, or unknown if it cannot infer it.