Refactor information models/API
CongressInfo is now two seperate tables: SenateInfo and HouseInfo. The SenateInfo table holds the date of the next session, whether it is currently in session, and the link to the live stream (if its in session it works). The HouseInfo table holds the same data, but right now only whether it is in session is implemented.